+++ ArduSub/failsafe.cpp
@@ -149,24 +149,6 @@ void Sub::failsafe_ekf_check()
     }
 }
 
-// Battery failsafe handler
-void Sub::handle_battery_failsafe(const char* type_str, const int8_t action)
-{
-    LOGGER_WRITE_ERROR(LogErrorSubsystem::FAILSAFE_BATT, LogErrorCode::FAILSAFE_OCCURRED);
-
-    switch((Failsafe_Action)action) {
-        case Failsafe_Action_Surface:
-            set_mode(Mode::Number::SURFACE, ModeReason::BATTERY_FAILSAFE);
-            break;
-        case Failsafe_Action_Disarm:
-            arming.disarm(AP_Arming::Method::BATTERYFAILSAFE);
-            break;
-        case Failsafe_Action_Warn:
-        case Failsafe_Action_None:
-            break;
-    }
-}
-
 // Make sure that we are receiving pilot input at an appropriate interval
 void Sub::failsafe_pilot_input_check()
 {
@@ -465,42 +447,6 @@ void Sub::failsafe_terrain_set_status(bool data_ok)
     }
 }
 
-// terrain failsafe action
-void Sub::failsafe_terrain_on_event()
-{
-    failsafe.terrain = true;
-    LOGGER_WRITE_ERROR(LogErrorSubsystem::FAILSAFE_TERRAIN, LogErrorCode::FAILSAFE_OCCURRED);
-
-    // If rangefinder is enabled, we can recover from this failsafe
-    if (!rangefinder_state.enabled || !sub.mode_auto.auto_terrain_recover_start()) {
-        failsafe_terrain_act();
-    }
-
-
-}
-
-// Recovery failed, take action
-void Sub::failsafe_terrain_act()
-{
-    switch (g.failsafe_terrain) {
-    case FS_TERRAIN_HOLD:
-        if (!set_mode(Mode::Number::POSHOLD, ModeReason::TERRAIN_FAILSAFE)) {
-            set_mode(Mode::Number::ALT_HOLD, ModeReason::TERRAIN_FAILSAFE);
-        }
-        AP_Notify::events.failsafe_mode_change = 1;
-        break;
-
-    case FS_TERRAIN_SURFACE:
-        set_mode(Mode::Number::SURFACE, ModeReason::TERRAIN_FAILSAFE);
-        AP_Notify::events.failsafe_mode_change = 1;
-        break;
-
-    case FS_TERRAIN_DISARM:
-    default:
-        arming.disarm(AP_Arming::Method::TERRAINFAILSAFE);
-    }
-}
-
 #if AP_SUB_RC_ENABLED
 void Sub::set_failsafe_radio(bool b)
 {
@@ -528,29 +474,4 @@ void Sub::set_failsafe_radio(bool b)
     }
 }
 
-// failsafe_radio_on_event - RC contact lost
-void Sub::failsafe_radio_on_event()
-{
-    LOGGER_WRITE_ERROR(LogErrorSubsystem::FAILSAFE_RADIO, LogErrorCode::FAILSAFE_OCCURRED);
-    gcs().send_text(MAV_SEVERITY_WARNING, "RC Failsafe");
-        switch(g.failsafe_throttle) {
-        case FS_THR_SURFACE:
-            set_mode(Mode::Number::SURFACE, ModeReason::RADIO_FAILSAFE);
-            break;
-        case FS_THR_WARN:
-            set_neutral_controls();
-            break;
-        case FS_THR_DISABLED:
-            break;
-    }    
-}
-
-// failsafe_radio_off event- respond to radio contact being regained
-void Sub::failsafe_radio_off_event()
-{
-    // no need to do anything except log the error as resolved
-    // user can now override roll, pitch, yaw and throttle and even use flight mode switch to restore previous flight mode
-    LOGGER_WRITE_ERROR(LogErrorSubsystem::FAILSAFE_RADIO, LogErrorCode::FAILSAFE_RESOLVED);
-    gcs().send_text(MAV_SEVERITY_WARNING, "Radio Failsafe Cleared");
-}
 #endif
